requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.1.0=gpu
absl-py=1.4.0=pypi_0
accelerate=0.20.3=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.5.0=py37h06a4308_0
argon2-cffi=20.1.0=py37h27cfd23_1
astor=0.8.1=py37h06a4308_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=py37h06a4308_0
asynctest=0.13.0=py_0
attrs=22.1.0=py37h06a4308_0
babel=2.11.0=py37h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py37h06a4308_0
blas=1.0=openblas
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py37h06a4308_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.1=pypi_0
cairo=1.16.0=hb05425b_5
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cma=3.3.0=pypi_0
cryptography=39.0.1=py37h9ce1e76_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cycler=0.11.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diffusers=0.17.1=pypi_0
dlib=19.24.2=pypi_0
eigen=3.4.0=h4bd325d_0
entrypoints=0.4=py37h06a4308_0
expat=2.4.9=h6a678d5_0
face-recognition=1.3.0=pypi_0
face-recognition-models=0.3.0=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.12.2=pypi_0
flatbuffers=23.5.26=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py37h5eee18b_0
fsspec=2023.1.0=pypi_0
gast=0.4.0=pypi_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.21.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pyhd3eb1b0_0
graphite2=1.3.14=h295c915_1
grpcio=1.56.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.1.0=nompi_py37h1e651dc_100
harfbuzz=4.3.0=hf52aaf7_1
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.15.1=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py37h06a4308_0
imageio=2.31.2=pypi_0
importlib-metadata=4.11.3=py37h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.15.2=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37h06a4308_1
jinja2=3.1.2=py37h06a4308_0
jpeg=9e=h5eee18b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=py37h06a4308_0
jupyter=1.0.0=py37h06a4308_8
jupyter_client=7.4.9=py37h06a4308_0
jupyter_console=6.4.4=py37h06a4308_0
jupyter_core=4.11.2=py37h06a4308_0
jupyter_server=1.23.4=py37h06a4308_0
jupyterlab=3.5.3=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.19.0=py37h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.11.0=pypi_0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py37_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libclang=16.0.0=pypi_0
libclang13=14.0.6=default_he11475f_1
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=15_linux64_openblas
libllvm14=14.0.6=hdb19cb5_3
libopenblas=0.3.20=pthreads_h78a6416_0
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.3=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=py37h06a4308_0
mistune=0.8.4=py37h14c3975_1001
mtcnn=0.1.1=pyhd8ed1ab_0
multidict=6.0.4=pypi_0
nbclassic=0.5.2=py37h06a4308_0
nbclient=0.5.13=py37h06a4308_0
nbconvert=6.4.4=py37h06a4308_0
nbformat=5.7.0=py37h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py37h06a4308_0
nettle=3.6=he412f7d_0
networkx=2.6.3=pypi_0
notebook=6.5.2=py37h06a4308_0
notebook-shim=0.2.2=py37h06a4308_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numpy=1.21.6=py37h976b520_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.27.8=pypi_0
opencv=4.6.0=py37h417a72b_3
opencv-python=4.8.0.74=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=22.0=py37h06a4308_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=22.3.1=py37h06a4308_0
pixman=0.40.0=h36c2ea0_0
pkgutil-resolve-name=1.3.10=py37h06a4308_0
ply=3.11=py37_0
prometheus_client=0.14.1=py37h06a4308_0
prompt-toolkit=3.0.36=py37h06a4308_0
prompt_toolkit=3.0.36=hd3eb1b0_0
protobuf=3.19.6=pypi_0
psutil=5.9.0=py37h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py37h06a4308_0
pyopenssl=23.0.0=py37h06a4308_0
pyparsing=3.1.0=pypi_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37_1
pytelegrambotapi=4.12.0=pypi_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py37h06a4308_0
python-version=0.0.2=pypi_0
python_abi=3.7=2_cp37m
pytz=2022.7=py37h06a4308_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=py37h540881e_4
pyzmq=23.2.0=py37h6a678d5_0
qt-main=5.15.2=h8373d8f_8
qt-webengine=5.15.9=hbbf29b9_6
qtconsole=5.4.0=py37h06a4308_0
qtpy=2.2.0=py37h06a4308_0
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
regex=2023.6.3=pypi_0
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.1=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.7.3=py37hf2a6cf1_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.6.3=py37h06a4308_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py37h06a4308_1
soupsieve=2.3.2.post1=py37h06a4308_0
sqlite=3.41.2=h5eee18b_0
telebot=0.0.5=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=py37h52d8a92_0
tensorboard-plugin-wit=1.8.1=py37h06a4308_0
tensorflow=2.11.0=pypi_0
tensorflow-base=2.1.0=gpu_py37h6c5654b_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-gpu=2.1.0=h0d30ee6_0
tensorflow-io-gcs-filesystem=0.32.0=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
tifffile=2021.11.2=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py37h06a4308_0
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=py37h5eee18b_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py37h06a4308_0
transformers=4.30.2=pypi_0
typing-extensions=4.3.0=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37h06a4308_4
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
widgetsnbextension=3.5.2=py37h06a4308_0
wrapt=1.15.0=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.9.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py37h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
